Move new tests to sycl/test/inline-asm
